fixing for update 2.2

+++ database/migrations/2019_05_08_053815_create_certificates_table.php
@@ -13,17 +13,20 @@ class CreateCertificatesTable extends Migration
      */
     public function up()
     {
-        Schema::create('certificates', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('name')->nullable();
-            $table->integer('user_id')->unsigned()->nullable();
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('course_id')->unsigned()->nullable();
-            $table->foreign('course_id')->references('id')->on('courses')->onDelete('cascade');
-            $table->text('url')->nullable();
-            $table->tinyInteger('status')->default(1)->comment('1-Generated 0-Not Generated');
-            $table->timestamps();
-        });
+        if(! Schema::hasTable('certificates')) {
+
+            Schema::create('certificates', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('name')->nullable();
+                $table->integer('user_id')->unsigned()->nullable();
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->integer('course_id')->unsigned()->nullable();
+                $table->foreign('course_id')->references('id')->on('courses')->onDelete('cascade');
+                $table->text('url')->nullable();
+                $table->tinyInteger('status')->default(1)->comment('1-Generated 0-Not Generated');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
